fix: fix crash when pressing back button in eco gauge
