--- app/src/main/java/com/b07/planetze/EcoGauge/EcoGaugeBreakdownCallback.java
@@ -1,7 +1,0 @@
-package com.b07.planetze.EcoGauge;
-
-import java.util.Map;
-
-public interface EcoGaugeBreakdownCallback {
-    Map<String, Double> getEmissionsByCategory();
-}
