--- app/src/main/java/com/b07/planetze/EcoGauge/EcoGaugeOverviewCallback.java
@@ -1,7 +1,0 @@
-package com.b07.planetze.EcoGauge;
-
-public interface EcoGaugeOverviewCallback {
-    public String getPeriod(String timePeriod);
-
-    public void setTimePeriod(String timePeriod);
-}
