--- app/src/main/java/com/b07/planetze/EcoGauge/EcoGaugeTrendsCallback.java
@@ -1,8 +1,0 @@
-package com.b07.planetze.EcoGauge;
-
-import com.jjoe64.graphview.series.DataPoint;
-import com.jjoe64.graphview.series.LineGraphSeries;
-
-public interface EcoGaugeTrendsCallback {
-    public LineGraphSeries<DataPoint> getDataPoints();
-}
