+++ app/src/main/java/com/b07/planetze/home/HomeScreenFragment.java
@@ -1,12 +1,9 @@
 package com.b07.planetze.home;
 
-import android.content.Context;
-import android.content.Intent;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.LinearLayout;
 import android.widget.TextView;
 import android.widget.Button;
 
@@ -16,10 +13,7 @@ import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentTransaction;
 import androidx.lifecycle.ViewModelProvider;
 
-import com.b07.planetze.EcoGauge.EcoGaugeActivity;
-import com.b07.planetze.EcoGauge.EcoGaugeScreen;
-import com.b07.planetze.EcoGauge.EcoGaugeScreenSwitch;
-import com.b07.planetze.MainActivity;
+import com.b07.planetze.ecogauge.EcoGaugeActivity;
 import com.b07.planetze.R;
 import com.b07.planetze.common.User;
 import com.b07.planetze.ecotracker.EcoTrackerActivity;
