+++ app/src/main/res/layout/activity_ecogauge.xml
@@ -5,7 +5,7 @@
 
 
     <FrameLayout
-        android:id="@+id/fragment_container"
+        android:id="@+id/ecogauge_fragment_container"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
