+++ app/src/main/res/layout/fragment_eco_gauge_breakdown.xml
@@ -6,7 +6,6 @@
 
 
     <LinearLayout
-        xmlns:android="http://schemas.android.com/apk/res/android"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:padding="10dp"
@@ -32,7 +31,7 @@
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_marginLeft="10dp"
-                android:text="Eco Gauge Breakdown"
+                android:text="Breakdown"
                 android:textSize="35sp"
                 android:textStyle="bold" />
         </LinearLayout>
