+++ app/src/main/res/layout/fragment_eco_gauge_trends.xml
@@ -4,13 +4,12 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".EcoGauge.EcoGaugeTrendsFragment">
+    tools:context=".ecogauge.EcoGaugeTrendsFragment">
 
     <!--line graph view where we will
         be displaying our data-->
 
     <LinearLayout
-        xmlns:android="http://schemas.android.com/apk/res/android"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:padding="10dp"
@@ -36,7 +35,7 @@
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_marginLeft="10dp"
-                android:text="Eco Gauge Trends"
+                android:text="Trends"
                 android:textSize="35sp"
                 android:textStyle="bold" />
         </LinearLayout>
